added event handler and split up the client.on events into the events folder
