+++ events/messageCreate.js
@@ -1,0 +1,21 @@
+module.exports = {
+    name: "messageCreate",
+    execute(interaction) {
+        /* // Ignore messages sent by the bot
+        if (interaction.author == client.user) return;
+
+        // Setting up the source channel
+        // and the target channel
+        if (interaction.channelId === '1019734223843774554') {
+            const targetChannel = client.channels.resolve('1021710965777117184');
+            targetChannel.send(interaction);
+
+        // This is the same thing
+        // but the other way around
+        } else if (interaction.channelId === '1021710965777117184') {
+            const targetChannel = client.channels.resolve('1019734223843774554');
+            targetChannel.send(interaction);
+        } */
+        return
+    }
+}
